Global variables fixes
%%% Removing the global variables

In matconvnet, options parsing takes two kind of inputs
- a (param, value) pair
- a struct, which can be converted by vl_argparse into a sequence of (param, values) pairs.

In particular, the cnn_train function is given options through the structures
- net.meta (often defined in the init script)
- opts.train (can be passed to cnn_mnist via the pair ('train', trainstruct)

Now, most of the options related to the solver are set in a structure 'train'
defined in the experiment file, and passed to cnn_mnist. The init file contains
the definition of the empty struct net.meta.trainOpts = struct();

%%% Changes in cnn_mnist.m (simplenn autoencoder)

- removed global variables solverfn, and initfn and the extra option to
cnn_train: (errorFunction,psnr), (solver, solver)...
The  remaining differences with the standard cnn_mnist.m are
- getBatch functions,
- data reshaping for fully connected,
- rescaling to [0,1],
- no subtracting the mean.

%%% Changes in cnn_mnist.m (dagnn with BN autoencoder)
- Same as above

Network and training options are now defined in a structure and passed as
options to cnn_mnist.
 - the training options are not in the init anymore.

%%% Added file Diff.m in dagnn for the wiring implementation with one BN